[balances-tx-payment-extension] tests work
