+++ balances-tx-payment-extension/Cargo.toml
@@ -15,6 +15,14 @@ pallet-asset-tx-payment = { version = "4.0.0-dev", git = "https://github.com/par
 pallet-transaction-payment = { version = "4.0.0-dev", git = "https://github.com/paritytech/substrate.git", default-features = false, branch = "master" }
 sp-runtime = { version = "6.0.0", git = "https://github.com/paritytech/substrate.git", default-features = false, branch = "master" }
 
+[dev-dependencies]
+approx = "0.5.1"
+codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false, features = ["derive"] }
+scale-info = { version = "2.0.1", default-features = false }
+sp-io = { version = "6.0.0", git = "https://github.com/paritytech/substrate.git", branch = "master" }
+test-utils = { path = "../test-utils" }
+
+
 [features]
 default = ["std"]
 std = [
