+++ balances-tx-payment-extension/src/lib.rs
@@ -1,6 +1,9 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 pub mod balance_conversion;
-
+#[cfg(test)]
+mod mock;
 #[cfg(test)]
 mod tests;
+
+pub use balance_conversion::*;
